switch back to the fancy sqrt algorithm
i overlooked the fact that we could delay the inversion indefinitely
by carrying x1d out as the z element of the point. that cuts out the last
inversion, which is more than worth the extra complexity and a few more
multiplications. the bulk of the cost is now the sqrt call.

           │  take5.txt   │               take6.txt                │
           │    sec/op    │   sec/op     vs base                   │
MapToCurve   11.501µ ± 2%   6.162µ ± 1%  -46.42% (p=0.000 n=10+30)